+++ companion/components/AuthTokenProvider.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { useAuth } from "../contexts/AuthContext";
-import { setAuthToken, setWebSession } from "../services/calcom";
+import { CalComAPIService } from "../services/calcom";
 
 export function AuthTokenProvider({ children }: { children: React.ReactNode }) {
   const { accessToken, isAuthenticated, isWebSession } = useAuth();
@@ -8,12 +8,10 @@ export function AuthTokenProvider({ children }: { children: React.ReactNode }) {
   useEffect(() => {
     if (isAuthenticated) {
       if (accessToken) {
-        setAuthToken(accessToken);
+        CalComAPIService.setAccessToken(accessToken);
       }
-      setWebSession(isWebSession);
     } else {
-      setAuthToken(null);
-      setWebSession(false);
+      CalComAPIService.clearAuth();
     }
   }, [isAuthenticated, accessToken, isWebSession]);
