+++ companion/components/event-type-detail/tabs/AvailabilityTab.tsx
@@ -16,7 +16,7 @@ interface AvailabilityTabProps {
   schedulesLoading: boolean;
   scheduleDetailsLoading: boolean;
   selectedScheduleDetails: any;
-  getDaySchedule: () => DaySchedule[];
+  getDaySchedules: () => DaySchedule[];
   formatTime: (time: string) => string;
   selectedTimezone: string;
 }
@@ -60,7 +60,7 @@ export function AvailabilityTab(props: AvailabilityTabProps) {
                 <Text className="text-sm italic text-[#8E8E93]">Loading schedule details...</Text>
               </View>
             ) : props.selectedScheduleDetails ? (
-              props.getDaySchedule().map((dayInfo, index) => (
+              props.getDaySchedules().map((dayInfo, index) => (
                 <View key={index} className="flex-row items-center justify-between py-4">
                   <Text
                     className={`ml-2 flex-1 text-[15px] font-medium text-[#333] ${
