+++ companion/contexts/AuthContext.tsx
@@ -36,6 +36,10 @@ interface AuthProviderProps {
 // Use the shared secure storage adapter
 const storage = secureStorage;
 
+const getErrorMessage = (error: unknown) =>
+  error instanceof Error ? error.message : String(error);
+const getErrorStack = (error: unknown) => (error instanceof Error ? error.stack : undefined);
+
 export function AuthProvider({ children }: AuthProviderProps) {
   const [isAuthenticated, setIsAuthenticated] = useState(false);
   const [accessToken, setAccessToken] = useState<string | null>(null);
@@ -143,7 +147,14 @@ export function AuthProvider({ children }: AuthProviderProps) {
           newRefreshToken || refreshToken || undefined
         );
       } catch (error) {
-        console.error("Failed to handle token refresh:", error);
+        const message = getErrorMessage(error);
+        console.error("Failed to handle token refresh", message);
+        if (__DEV__) {
+          console.debug("[AuthContext] token refresh handler failed", {
+            message,
+            stack: getErrorStack(error),
+          });
+        }
         await logout();
       }
     };
@@ -213,7 +224,14 @@ export function AuthProvider({ children }: AuthProviderProps) {
         handleWebSessionAuth();
       }
     } catch (error) {
-      console.error("Failed to check auth state:", error);
+      const message = getErrorMessage(error);
+      console.error("Failed to check auth state", message);
+      if (__DEV__) {
+        console.debug("[AuthContext] checkAuthState failed", {
+          message,
+          stack: getErrorStack(error),
+        });
+      }
     } finally {
       setLoading(false);
     }
@@ -234,7 +252,14 @@ export function AuthProvider({ children }: AuthProviderProps) {
         await setupAfterLogin(tokens.accessToken, tokens.refreshToken);
       }
     } catch (error) {
-      console.error("Failed to login from web session:", error);
+      const message = getErrorMessage(error);
+      console.error("Failed to login from web session", message);
+      if (__DEV__) {
+        console.debug("[AuthContext] loginFromWebSession failed", {
+          message,
+          stack: getErrorStack(error),
+        });
+      }
       throw error;
     }
   };
@@ -267,7 +292,14 @@ export function AuthProvider({ children }: AuthProviderProps) {
       // Clear PKCE parameters
       oauthService.clearPKCEParams();
     } catch (error) {
-      console.error("OAuth login failed:", error);
+      const message = getErrorMessage(error);
+      console.error("OAuth login failed", message);
+      if (__DEV__) {
+        console.debug("[AuthContext] loginWithOAuth failed", {
+          message,
+          stack: getErrorStack(error),
+        });
+      }
       throw error;
     } finally {
       setLoading(false);
@@ -279,7 +311,11 @@ export function AuthProvider({ children }: AuthProviderProps) {
       await clearAuth();
       resetAuthState();
     } catch (error) {
-      console.error("Failed to logout:", error);
+      const message = getErrorMessage(error);
+      console.error("Failed to logout", message);
+      if (__DEV__) {
+        console.debug("[AuthContext] logout failed", { message, stack: getErrorStack(error) });
+      }
     }
   };
