+++ companion/hooks/useBookings.ts
@@ -108,7 +108,7 @@ export function useCancelBooking() {
       });
     },
     onError: (error) => {
-      console.error("Failed to cancel booking:", error);
+      console.error("Failed to cancel booking");
     },
   });
 }
@@ -152,7 +152,7 @@ export function useRescheduleBooking() {
       });
     },
     onError: (error) => {
-      console.error("Failed to reschedule booking:", error);
+      console.error("Failed to reschedule booking");
     },
   });
 }
@@ -184,7 +184,7 @@ export function useConfirmBooking() {
       });
     },
     onError: (error) => {
-      console.error("Failed to confirm booking:", error);
+      console.error("Failed to confirm booking");
     },
   });
 }
@@ -217,7 +217,7 @@ export function useDeclineBooking() {
       });
     },
     onError: (error) => {
-      console.error("Failed to decline booking:", error);
+      console.error("Failed to decline booking");
     },
   });
 }
