+++ companion/hooks/useEventTypes.ts
@@ -93,7 +93,12 @@ export function useCreateEventType() {
       queryClient.setQueryData(queryKeys.eventTypes.detail(newEventType.id), newEventType);
     },
     onError: (error) => {
-      console.error("Failed to create event type:", error);
+      console.error("Failed to create event type");
+      if (__DEV__) {
+        const message = error instanceof Error ? error.message : String(error);
+        const stack = error instanceof Error ? error.stack : undefined;
+        console.debug("[useCreateEventType] failed", { message, stack });
+      }
     },
   });
 }
@@ -127,7 +132,12 @@ export function useUpdateEventType() {
       queryClient.setQueryData(queryKeys.eventTypes.detail(variables.id), updatedEventType);
     },
     onError: (error) => {
-      console.error("Failed to update event type:", error);
+      console.error("Failed to update event type");
+      if (__DEV__) {
+        const message = error instanceof Error ? error.message : String(error);
+        const stack = error instanceof Error ? error.stack : undefined;
+        console.debug("[useUpdateEventType] failed", { message, stack });
+      }
     },
   });
 }
@@ -173,7 +183,12 @@ export function useDeleteEventType() {
       if (context?.previousEventTypes) {
         queryClient.setQueryData(queryKeys.eventTypes.lists(), context.previousEventTypes);
       }
-      console.error("Failed to delete event type:", error);
+      console.error("Failed to delete event type");
+      if (__DEV__) {
+        const message = error instanceof Error ? error.message : String(error);
+        const stack = error instanceof Error ? error.stack : undefined;
+        console.debug("[useDeleteEventType] failed", { message, stack });
+      }
     },
     onSettled: () => {
       // Always refetch after error or success
@@ -233,7 +248,12 @@ export function useDuplicateEventType() {
       queryClient.invalidateQueries({ queryKey: queryKeys.eventTypes.lists() });
     },
     onError: (error) => {
-      console.error("Failed to duplicate event type:", error);
+      console.error("Failed to duplicate event type");
+      if (__DEV__) {
+        const message = error instanceof Error ? error.message : String(error);
+        const stack = error instanceof Error ? error.stack : undefined;
+        console.debug("[useDuplicateEventType] failed", { message, stack });
+      }
     },
   });
 }
