+++ companion/package.json
@@ -1,13 +1,14 @@
 {
   "name": "cal-companion",
   "displayName": "Cal.com Companion",
-  "version": "1.7.4",
+  "version": "1.7.5",
   "main": "index.js",
   "scripts": {
     "start": "expo start",
     "android": "expo run:android",
     "ios": "expo run:ios",
     "web": "expo start --web",
+    "typecheck": "tsc -p tsconfig.json --noEmit",
     "ext": "wxt",
     "ext:build": "wxt build",
     "ext:build-prod": "BUILD_FOR_STORE=true wxt build",
