+++ companion/utils/gmailGoogleChipParser.ts
@@ -149,7 +149,12 @@ export function parseGoogleChip(chipElement: HTMLElement): ParsedGoogleChip | nu
       detectedDuration,
     };
   } catch (error) {
-    console.error("Error parsing Google chip:", error);
+    console.error("Error parsing Google chip");
+    if (__DEV__) {
+      const message = error instanceof Error ? error.message : String(error);
+      const stack = error instanceof Error ? error.stack : undefined;
+      console.debug("[gmailGoogleChipParser] parse failed", { message, stack });
+    }
     return null;
   }
 }
