+++ companion/wxt.config.ts
@@ -58,7 +58,7 @@ export default defineConfig({
   outDir: ".output",
   manifest: {
     name: "Cal.com Companion",
-    version: "1.7.4",
+    version: "1.7.5",
     description: "Your calendar companion for quick booking and scheduling",
     permissions: ["activeTab", "storage", "identity"],
     host_permissions: [
@@ -68,7 +68,7 @@ export default defineConfig({
       "https://mail.google.com/*",
       "https://calendar.google.com/*",
       // Include localhost permission for dev builds (needed for iframe to load)
-      ...(!isBuildForStore ? ["http://localhost:*/*"] : []),
+      // ...(!isBuildForStore ? ["http://localhost:*/*"] : []),
     ],
     content_security_policy: {
       extension_pages: !isBuildForStore
